local_backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
# extras for testing
pytest
requests
reportlab
pillow
